+++ ydb/core/external_sources/s3/ut/ya.make
@@ -52,9 +52,7 @@ IF (OPENSOURCE)
 ENDIF()
 
 SRCS(
-    common.cpp
     s3_aws_credentials_ut.cpp
-    s3_insert_ut.cpp
 )
 
 PEERDIR(
